util-requirements.txt
# env_ops
boto3
python-dotenv

# pdf_ops
PyPDF2